+++ packages/@codotype/types/lib/configuration-option-types.js
@@ -1,0 +1,9 @@
+// ConfigurationOption Type IDs
+
+const OPTION_TYPE_STRING_SELECT = 'STRING_SELECT';
+
+// // // //
+
+module.exports = {
+  OPTION_TYPE_STRING_SELECT
+}